removed infrastructure-metrics module
