+++ infrastructures/pom.xml
@@ -27,6 +27,5 @@
     <modules>
         <module>openshift</module>
         <module>kubernetes</module>
-        <module>infrastructure-metrics</module>
     </modules>
 </project>
